dialects: (riscv) split into multiple files
I aimed to preserve the existing API, hence the re-exports, which will be deprecated and removed in the future.
