+++ xdsl/dialects/riscv/ops.py
@@ -1,23 +1,13 @@
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from collections.abc import Set as AbstractSet
-from io import StringIO
-from typing import IO, ClassVar, Generic, Literal, TypeAlias, cast
-
-from typing_extensions import Self, TypeVar
+from typing import ClassVar, cast
 
 from xdsl.backend.assembly_printer import (
     AssemblyPrintable,
     AssemblyPrinter,
-    OneLineAssemblyPrintable,
 )
-from xdsl.backend.register_allocatable import (
-    HasRegisterConstraints,
-    RegisterConstraints,
-)
-from xdsl.backend.register_type import RegisterAllocatedMemoryEffect, RegisterType
 from xdsl.dialects.builtin import (
     I32,
     ArrayAttr,
@@ -25,22 +15,14 @@ from xdsl.dialects.builtin import (
     IndexType,
     IntegerAttr,
     IntegerType,
-    ModuleOp,
-    NoneAttr,
-    Signedness,
     StringAttr,
-    UnitAttr,
     i32,
 )
-from xdsl.dialects.utils import FastMathAttrBase, FastMathFlag
 from xdsl.interfaces import ConstantLikeInterface
 from xdsl.ir import (
     Attribute,
     Block,
-    Data,
     Dialect,
-    Operation,
-    OpResult,
     Region,
     SSAValue,
 )
@@ -52,9 +34,7 @@ from xdsl.irdl import (
     RangeOf,
     VarOpResult,
     attr_def,
-    irdl_attr_definition,
     irdl_op_definition,
-    operand_def,
     opt_attr_def,
     opt_prop_def,
     prop_def,
@@ -64,7 +44,7 @@ from xdsl.irdl import (
     var_operand_def,
     var_result_def,
 )
-from xdsl.parser import AttrParser, Parser, UnresolvedOperand
+from xdsl.parser import Parser
 from xdsl.pattern_rewriter import RewritePattern
 from xdsl.printer import Printer
 from xdsl.traits import (
@@ -76,1372 +56,47 @@ from xdsl.traits import (
 )
 from xdsl.utils.exceptions import VerifyException
 
-
-def is_non_zero(reg: IntRegisterType) -> bool:
-    """
-    Returns True if the register is allocated, and is not the x0/ZERO register.
-    """
-    return (
-        reg.is_allocated and not isinstance(reg.index, NoneAttr) and reg.index.data != 0
-    )
-
-
-@irdl_attr_definition
-class FastMathFlagsAttr(FastMathAttrBase):
-    """
-    riscv.fastmath is a mirror of LLVMs fastmath flags.
-    """
-
-    name = "riscv.fastmath"
-
-    def __init__(self, flags: None | Sequence[FastMathFlag] | Literal["none", "fast"]):
-        # irdl_attr_definition defines an __init__ if none is defined, so we need to
-        # explicitely define one here.
-        super().__init__(flags)
-
-
-class RISCVRegisterType(RegisterType):
-    """
-    A RISC-V register type.
-    """
-
-    @classmethod
-    @abstractmethod
-    def a_register(cls, index: int) -> Self:
-        raise NotImplementedError()
-
-
-_RV32I_ABI_INDEX_BY_NAME = {
-    "zero": 0,
-    "ra": 1,
-    "sp": 2,
-    "gp": 3,
-    "tp": 4,
-    "t0": 5,
-    "t1": 6,
-    "t2": 7,
-    "fp": 8,
-    "s0": 8,
-    "s1": 9,
-    "a0": 10,
-    "a1": 11,
-    "a2": 12,
-    "a3": 13,
-    "a4": 14,
-    "a5": 15,
-    "a6": 16,
-    "a7": 17,
-    "s2": 18,
-    "s3": 19,
-    "s4": 20,
-    "s5": 21,
-    "s6": 22,
-    "s7": 23,
-    "s8": 24,
-    "s9": 25,
-    "s10": 26,
-    "s11": 27,
-    "t3": 28,
-    "t4": 29,
-    "t5": 30,
-    "t6": 31,
-}
-_RV32I_X_INDEX_BY_NAME = {f"x{i}": i for i in range(32)}
-RV32I_INDEX_BY_NAME = _RV32I_X_INDEX_BY_NAME | _RV32I_ABI_INDEX_BY_NAME
-
-
-@irdl_attr_definition
-class IntRegisterType(RISCVRegisterType):
-    """
-    A RISC-V register type.
-    """
-
-    name = "riscv.reg"
-
-    @classmethod
-    def index_by_name(cls) -> dict[str, int]:
-        return RV32I_INDEX_BY_NAME
-
-    @classmethod
-    def a_register(cls, index: int) -> IntRegisterType:
-        return Registers.A[index]
-
-    @classmethod
-    def infinite_register_prefix(cls):
-        return "j_"
-
-    # This class variable is created and exclusively accessed in `abi_name_by_index`.
-    # _ALLOCATABLE_REGISTERS: ClassVar[tuple[IntRegisterType, ...]]
-
-    @classmethod
-    def allocatable_registers(cls):
-        if not hasattr(cls, "_ALLOCATABLE_REGISTERS"):
-            cls._ALLOCATABLE_REGISTERS = (*Registers.T, *Registers.A)
-        return cls._ALLOCATABLE_REGISTERS
-
-
-_RV32F_ABI_INDEX_BY_NAME = {
-    "ft0": 0,
-    "ft1": 1,
-    "ft2": 2,
-    "ft3": 3,
-    "ft4": 4,
-    "ft5": 5,
-    "ft6": 6,
-    "ft7": 7,
-    "fs0": 8,
-    "fs1": 9,
-    "fa0": 10,
-    "fa1": 11,
-    "fa2": 12,
-    "fa3": 13,
-    "fa4": 14,
-    "fa5": 15,
-    "fa6": 16,
-    "fa7": 17,
-    "fs2": 18,
-    "fs3": 19,
-    "fs4": 20,
-    "fs5": 21,
-    "fs6": 22,
-    "fs7": 23,
-    "fs8": 24,
-    "fs9": 25,
-    "fs10": 26,
-    "fs11": 27,
-    "ft8": 28,
-    "ft9": 29,
-    "ft10": 30,
-    "ft11": 31,
-}
-_RV32F_F_INDEX_BY_NAME = {f"f{i}": i for i in range(32)}
-RV32F_INDEX_BY_NAME = _RV32F_F_INDEX_BY_NAME | _RV32F_ABI_INDEX_BY_NAME
-
-
-@irdl_attr_definition
-class FloatRegisterType(RISCVRegisterType):
-    """
-    A RISC-V register type.
-    """
-
-    name = "riscv.freg"
-
-    @classmethod
-    def index_by_name(cls) -> dict[str, int]:
-        return RV32F_INDEX_BY_NAME
-
-    @classmethod
-    def a_register(cls, index: int) -> FloatRegisterType:
-        return Registers.FA[index]
-
-    @classmethod
-    def infinite_register_prefix(cls):
-        return "fj_"
-
-    # This class variable is created and exclusively accessed in `abi_name_by_index`.
-    # _ALLOCATABLE_REGISTERS: ClassVar[tuple[FloatRegisterType, ...]]
-
-    @classmethod
-    def allocatable_registers(cls):
-        if not hasattr(cls, "_ALLOCATABLE_REGISTERS"):
-            cls._ALLOCATABLE_REGISTERS = (*Registers.FT, *Registers.FA)
-        return cls._ALLOCATABLE_REGISTERS
-
-
-RDInvT = TypeVar("RDInvT", bound=RISCVRegisterType)
-RSInvT = TypeVar("RSInvT", bound=RISCVRegisterType)
-RS1InvT = TypeVar("RS1InvT", bound=RISCVRegisterType)
-RS2InvT = TypeVar("RS2InvT", bound=RISCVRegisterType)
-
-
-class Registers(ABC):
-    """Namespace for named register constants."""
-
-    UNALLOCATED_INT = IntRegisterType.unallocated()
-    ZERO = IntRegisterType.from_name("zero")
-    RA = IntRegisterType.from_name("ra")
-    SP = IntRegisterType.from_name("sp")
-    GP = IntRegisterType.from_name("gp")
-    TP = IntRegisterType.from_name("tp")
-    T0 = IntRegisterType.from_name("t0")
-    T1 = IntRegisterType.from_name("t1")
-    T2 = IntRegisterType.from_name("t2")
-    FP = IntRegisterType.from_name("fp")
-    S0 = IntRegisterType.from_name("s0")
-    S1 = IntRegisterType.from_name("s1")
-    A0 = IntRegisterType.from_name("a0")
-    A1 = IntRegisterType.from_name("a1")
-    A2 = IntRegisterType.from_name("a2")
-    A3 = IntRegisterType.from_name("a3")
-    A4 = IntRegisterType.from_name("a4")
-    A5 = IntRegisterType.from_name("a5")
-    A6 = IntRegisterType.from_name("a6")
-    A7 = IntRegisterType.from_name("a7")
-    S2 = IntRegisterType.from_name("s2")
-    S3 = IntRegisterType.from_name("s3")
-    S4 = IntRegisterType.from_name("s4")
-    S5 = IntRegisterType.from_name("s5")
-    S6 = IntRegisterType.from_name("s6")
-    S7 = IntRegisterType.from_name("s7")
-    S8 = IntRegisterType.from_name("s8")
-    S9 = IntRegisterType.from_name("s9")
-    S10 = IntRegisterType.from_name("s10")
-    S11 = IntRegisterType.from_name("s11")
-    T3 = IntRegisterType.from_name("t3")
-    T4 = IntRegisterType.from_name("t4")
-    T5 = IntRegisterType.from_name("t5")
-    T6 = IntRegisterType.from_name("t6")
-
-    UNALLOCATED_FLOAT = FloatRegisterType.unallocated()
-    FT0 = FloatRegisterType.from_name("ft0")
-    FT1 = FloatRegisterType.from_name("ft1")
-    FT2 = FloatRegisterType.from_name("ft2")
-    FT3 = FloatRegisterType.from_name("ft3")
-    FT4 = FloatRegisterType.from_name("ft4")
-    FT5 = FloatRegisterType.from_name("ft5")
-    FT6 = FloatRegisterType.from_name("ft6")
-    FT7 = FloatRegisterType.from_name("ft7")
-    FS0 = FloatRegisterType.from_name("fs0")
-    FS1 = FloatRegisterType.from_name("fs1")
-    FA0 = FloatRegisterType.from_name("fa0")
-    FA1 = FloatRegisterType.from_name("fa1")
-    FA2 = FloatRegisterType.from_name("fa2")
-    FA3 = FloatRegisterType.from_name("fa3")
-    FA4 = FloatRegisterType.from_name("fa4")
-    FA5 = FloatRegisterType.from_name("fa5")
-    FA6 = FloatRegisterType.from_name("fa6")
-    FA7 = FloatRegisterType.from_name("fa7")
-    FS2 = FloatRegisterType.from_name("fs2")
-    FS3 = FloatRegisterType.from_name("fs3")
-    FS4 = FloatRegisterType.from_name("fs4")
-    FS5 = FloatRegisterType.from_name("fs5")
-    FS6 = FloatRegisterType.from_name("fs6")
-    FS7 = FloatRegisterType.from_name("fs7")
-    FS8 = FloatRegisterType.from_name("fs8")
-    FS9 = FloatRegisterType.from_name("fs9")
-    FS10 = FloatRegisterType.from_name("fs10")
-    FS11 = FloatRegisterType.from_name("fs11")
-    FT8 = FloatRegisterType.from_name("ft8")
-    FT9 = FloatRegisterType.from_name("ft9")
-    FT10 = FloatRegisterType.from_name("ft10")
-    FT11 = FloatRegisterType.from_name("ft11")
-
-    # register classes:
-
-    A = (A0, A1, A2, A3, A4, A5, A6, A7)
-    T = (T0, T1, T2, T3, T4, T5, T6)
-    S = (S0, S1, S2, S3, S4, S5, S6, S7, S8, S9, S10, S11)
-
-    FA = (FA0, FA1, FA2, FA3, FA4, FA5, FA6, FA7)
-    FT = (FT0, FT1, FT2, FT3, FT4, FT5, FT6, FT7, FT8, FT9, FT10, FT11)
-    FS = (FS0, FS1, FS2, FS3, FS4, FS5, FS6, FS7, FS8, FS9, FS10, FS11)
-
-
-UI5: TypeAlias = IntegerType[Literal[5], Literal[Signedness.UNSIGNED]]
-SI20: TypeAlias = IntegerType[Literal[20], Literal[Signedness.SIGNED]]
-SI12: TypeAlias = IntegerType[Literal[12], Literal[Signedness.SIGNED]]
-I12: TypeAlias = IntegerType[Literal[12], Literal[Signedness.SIGNLESS]]
-I20: TypeAlias = IntegerType[Literal[20], Literal[Signedness.SIGNLESS]]
-ui5: UI5 = IntegerType(5, Signedness.UNSIGNED)
-si20: SI20 = IntegerType(20, Signedness.SIGNED)
-si12: SI12 = IntegerType(12, Signedness.SIGNED)
-i12: I12 = IntegerType(12, Signedness.SIGNLESS)
-i20: I20 = IntegerType(20, Signedness.SIGNLESS)
-
-
-@irdl_attr_definition
-class LabelAttr(Data[str]):
-    name = "riscv.label"
-
-    @classmethod
-    def parse_parameter(cls, parser: AttrParser) -> str:
-        with parser.in_angle_brackets():
-            return parser.parse_str_literal()
-
-    def print_parameter(self, printer: Printer) -> None:
-        with printer.in_angle_brackets():
-            printer.print_string_literal(self.data)
-
-
-class RISCVAsmOperation(IRDLOperation, OneLineAssemblyPrintable, ABC):
-    """
-    Base class for operations that can be a part of RISC-V assembly printing.
-    """
-
-
-class RISCVRegallocOperation(HasRegisterConstraints, IRDLOperation, ABC):
-    """
-    Base class for operations that can take part in register allocation.
-    """
-
-    def get_register_constraints(self) -> RegisterConstraints:
-        # The default register constraints are that all operands are "in", and all
-        # results are "out" registers.
-        # If some registers are "inout" then this function must be overridden.
-        return RegisterConstraints(self.operands, self.results, ())
-
-
-class RISCVCustomFormatOperation(IRDLOperation, ABC):
-    """
-    Base class for RISC-V operations that specialize their custom format.
-    """
-
-    @classmethod
-    def parse(cls, parser: Parser) -> Self:
-        args = cls.parse_unresolved_operands(parser)
-        custom_attributes = cls.custom_parse_attributes(parser)
-        remaining_attributes = parser.parse_optional_attr_dict()
-        # TODO ensure distinct keys for attributes
-        attributes = custom_attributes | remaining_attributes
-        regions = parser.parse_region_list()
-        pos = parser.pos
-        operand_types, result_types = cls.parse_op_type(parser)
-        operands = parser.resolve_operands(args, operand_types, pos)
-        return cls.create(
-            operands=operands,
-            result_types=result_types,
-            attributes=attributes,
-            regions=regions,
-        )
-
-    @classmethod
-    def parse_unresolved_operands(cls, parser: Parser) -> list[UnresolvedOperand]:
-        """
-        Parse a list of comma separated unresolved operands.
-
-        Notice that this method will consume trailing comma.
-        """
-        if operand := parser.parse_optional_unresolved_operand():
-            operands = [operand]
-            while parser.parse_optional_punctuation(",") and (
-                operand := parser.parse_optional_unresolved_operand()
-            ):
-                operands.append(operand)
-            return operands
-        return []
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        """
-        Parse attributes with custom syntax. Subclasses may override this method.
-        """
-        return parser.parse_optional_attr_dict()
-
-    @classmethod
-    def parse_op_type(
-        cls, parser: Parser
-    ) -> tuple[Sequence[Attribute], Sequence[Attribute]]:
-        parser.parse_punctuation(":")
-        func_type = parser.parse_function_type()
-        return func_type.inputs.data, func_type.outputs.data
-
-    def print(self, printer: Printer) -> None:
-        if self.operands:
-            printer.print_string(" ")
-            printer.print_list(self.operands, printer.print_operand)
-        printed_attributes = self.custom_print_attributes(printer)
-        unprinted_attributes = {
-            name: attr
-            for name, attr in self.attributes.items()
-            if name not in printed_attributes
-        }
-        printer.print_op_attributes(unprinted_attributes)
-        printer.print_regions(self.regions)
-        self.print_op_type(printer)
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        """
-        Print attributes with custom syntax. Return the names of the attributes printed. Subclasses may override this method.
-        """
-        printer.print_op_attributes(self.attributes)
-        return self.attributes.keys()
-
-    def print_op_type(self, printer: Printer) -> None:
-        printer.print_string(" : ")
-        printer.print_operation_type(self)
-
-
-AssemblyInstructionArg: TypeAlias = (
-    IntegerAttr | LabelAttr | SSAValue | RegisterType | str
+from .abstract_ops import (
+    AssemblyInstructionArg,
+    CsrBitwiseImmOperation,
+    CsrBitwiseOperation,
+    CsrReadWriteImmOperation,
+    CsrReadWriteOperation,
+    GetAnyRegisterOperation,
+    NullaryOperation,
+    RdImmIntegerOperation,
+    RdImmJumpOperation,
+    RdRsFloatOperation,
+    RdRsImmFloatOperation,
+    RdRsImmIntegerOperation,
+    RdRsImmJumpOperation,
+    RdRsImmShiftOperation,
+    RdRsIntegerOperation,
+    RdRsRsFloatFloatIntegerOperationWithFastMath,
+    RdRsRsFloatOperation,
+    RdRsRsFloatOperationWithFastMath,
+    RdRsRsIntegerOperation,
+    RdRsRsRsFloatOperation,
+    RISCVAsmOperation,
+    RISCVCustomFormatOperation,
+    RISCVInstruction,
+    RISCVRegallocOperation,
+    RsRsImmFloatOperation,
+    RsRsImmIntegerOperation,
+    RsRsOffIntegerOperation,
+    assembly_arg_str,
 )
-
-
-class RISCVInstruction(RISCVAsmOperation, RISCVRegallocOperation, ABC):
-    """
-    Base class for operations that can be a part of RISC-V assembly printing. Must
-    represent an instruction in the RISC-V instruction set, and have the following format:
-
-    name arg0, arg1, arg2           # comment
-
-    The name of the operation will be used as the RISC-V assembly instruction name.
-    """
-
-    comment = opt_attr_def(StringAttr)
-    """
-    An optional comment that will be printed along with the instruction.
-    """
-
-    @abstractmethod
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg | None, ...]:
-        """
-        The arguments to the instruction, in the order they should be printed in the
-        assembly.
-        """
-        raise NotImplementedError()
-
-    def assembly_instruction_name(self) -> str:
-        """
-        By default, the name of the instruction is the same as the name of the operation.
-        """
-
-        return Dialect.split_name(self.name)[1]
-
-    def assembly_line(self) -> str | None:
-        # default assembly code generator
-        instruction_name = self.assembly_instruction_name()
-        arg_str = ", ".join(
-            _assembly_arg_str(arg)
-            for arg in self.assembly_line_args()
-            if arg is not None
-        )
-        return AssemblyPrinter.assembly_line(instruction_name, arg_str, self.comment)
-
-
-# region Assembly printing
-
-
-def _assembly_arg_str(arg: AssemblyInstructionArg) -> str:
-    if isinstance(arg, SSAValue):
-        if not isinstance(t := arg.type, RegisterType):
-            raise ValueError(f"Unexpected register type {t}")
-        return t.register_name.data
-    elif isinstance(arg, IntegerAttr):
-        return f"{arg.value.data}"
-    elif isinstance(arg, LabelAttr):
-        return arg.data
-    elif isinstance(arg, RegisterType):
-        return arg.register_name.data
-
-    return arg
-
-
-def print_assembly(module: ModuleOp, output: IO[str]) -> None:
-    printer = AssemblyPrinter(stream=output)
-    printer.print_module(module)
-
-
-def riscv_code(module: ModuleOp) -> str:
-    stream = StringIO()
-    print_assembly(module, stream)
-    return stream.getvalue()
-
-
-# endregion
-
-# region Base Operation classes
-
-
-class RdRsRsOperation(
-    RISCVCustomFormatOperation, RISCVInstruction, ABC, Generic[RDInvT, RS1InvT, RS2InvT]
-):
-    """
-    A base class for RISC-V operations that have one destination register, and two source
-    registers.
-
-    This is called R-Type in the RISC-V specification.
-    """
-
-    rd: OpResult[RDInvT] = result_def(RDInvT)
-    rs1 = operand_def(RS1InvT)
-    rs2 = operand_def(RS2InvT)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        *,
-        rd: RDInvT = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            operands=[rs1, rs2],
-            attributes={
-                "comment": comment,
-            },
-            result_types=[rd],
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.rs1, self.rs2
-
-
-class RdRsRsIntegerOperation(
-    RdRsRsOperation[IntRegisterType, RS1InvT, RS2InvT], ABC, Generic[RS1InvT, RS2InvT]
-):
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        *,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        super().__init__(rs1, rs2, rd=rd, comment=comment)
-
-
-class RdRsRsFloatOperation(
-    RdRsRsOperation[FloatRegisterType, RS1InvT, RS2InvT], ABC, Generic[RS1InvT, RS2InvT]
-):
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        *,
-        rd: FloatRegisterType = Registers.UNALLOCATED_FLOAT,
-        comment: str | StringAttr | None = None,
-    ):
-        super().__init__(rs1, rs2, rd=rd, comment=comment)
-
-
-class RdRsRsFloatOperationWithFastMath(
-    RISCVCustomFormatOperation, RISCVInstruction, ABC
-):
-    """
-    A base class for RISC-V operations that have one destination floating-point register,
-    and two source floating-point registers and can be annotated with fastmath flags.
-
-    This is called R-Type in the RISC-V specification.
-    """
-
-    rd = result_def(FloatRegisterType)
-    rs1 = operand_def(FloatRegisterType)
-    rs2 = operand_def(FloatRegisterType)
-    fastmath = opt_attr_def(FastMathFlagsAttr)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        *,
-        rd: FloatRegisterType = Registers.UNALLOCATED_FLOAT,
-        fastmath: FastMathFlagsAttr | None = None,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            operands=[rs1, rs2],
-            attributes={
-                "fastmath": fastmath,
-                "comment": comment,
-            },
-            result_types=[rd],
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.rs1, self.rs2
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        flags = FastMathFlagsAttr("none")
-        if parser.parse_optional_keyword("fastmath") is not None:
-            flags = FastMathFlagsAttr(FastMathFlagsAttr.parse_parameter(parser))
-        attributes["fastmath"] = flags
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        if self.fastmath is not None and self.fastmath != FastMathFlagsAttr("none"):
-            printer.print_string(" fastmath")
-            self.fastmath.print_parameter(printer)
-        return {"fastmath"}
-
-
-class RdImmIntegerOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations that have one destination register, and one
-    immediate operand (e.g. U-Type and J-Type instructions in the RISC-V spec).
-    """
-
-    rd = result_def(IntRegisterType)
-    immediate = attr_def(IntegerAttr[I20] | LabelAttr)
-
-    def __init__(
-        self,
-        immediate: int | IntegerAttr | str | LabelAttr,
-        *,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(immediate, int):
-            immediate = IntegerAttr(immediate, i20)
-        elif isinstance(immediate, str):
-            immediate = LabelAttr(immediate)
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            result_types=[rd],
-            attributes={
-                "immediate": immediate,
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["immediate"] = parse_immediate_value(parser, i20)
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(" ")
-        print_immediate_value(printer, self.immediate)
-        return {"immediate"}
-
-
-class RdImmJumpOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    In the RISC-V spec, this is the same as `RdImmOperation`. For jumps, the `rd` register
-    is neither an operand, because the stored value is overwritten, nor a result value,
-    because the value in `rd` is not defined after the jump back. So the `rd` makes the
-    most sense as an attribute.
-    """
-
-    rd = opt_attr_def(IntRegisterType)
-    """
-    The rd register here is not a register storing the result, rather the register where
-    the program counter is stored before jumping.
-    """
-    immediate = attr_def(IntegerAttr[SI20] | LabelAttr)
-
-    def __init__(
-        self,
-        immediate: int | IntegerAttr[SI20] | str | LabelAttr,
-        *,
-        rd: IntRegisterType | None = None,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(immediate, int):
-            immediate = IntegerAttr(immediate, si20)
-        elif isinstance(immediate, str):
-            immediate = LabelAttr(immediate)
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            attributes={
-                "immediate": immediate,
-                "rd": rd,
-                "comment": comment,
-            }
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg | None, ...]:
-        return self.rd, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["immediate"] = parse_immediate_value(parser, si20)
-        if parser.parse_optional_punctuation(","):
-            attributes["rd"] = parser.parse_attribute()
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(" ")
-        print_immediate_value(printer, self.immediate)
-        if self.rd is not None:
-            printer.print_string(", ")
-            printer.print_attribute(self.rd)
-        return {"immediate", "rd"}
-
-    def print_op_type(self, printer: Printer) -> None:
-        return
-
-    @classmethod
-    def parse_op_type(
-        cls, parser: Parser
-    ) -> tuple[Sequence[Attribute], Sequence[Attribute]]:
-        return (), ()
-
-
-class RdRsImmIntegerOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations that have one destination register, one source
-    register and one immediate operand.
-
-    This is called I-Type in the RISC-V specification.
-    """
-
-    rd = result_def(IntRegisterType)
-    rs1 = operand_def(IntRegisterType)
-    immediate = attr_def(IntegerAttr[SI12] | LabelAttr)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        immediate: int | IntegerAttr[SI12] | str | LabelAttr,
-        *,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(immediate, int):
-            immediate = IntegerAttr(immediate, si12)
-        elif isinstance(immediate, str):
-            immediate = LabelAttr(immediate)
-
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            operands=[rs1],
-            result_types=[rd],
-            attributes={
-                "immediate": immediate,
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.rs1, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["immediate"] = parse_immediate_value(parser, si12)
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(", ")
-        print_immediate_value(printer, self.immediate)
-        return {"immediate"}
-
-
-class RdRsImmShiftOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations that have one destination register, one source
-    register and one immediate operand.
-
-    This is called I-Type in the RISC-V specification.
-
-    Shifts by a constant are encoded as a specialization of the I-type format.
-    The shift amount is encoded in the lower 5 bits of the I-immediate field for RV32
-
-    For RV32I, SLLI, SRLI, and SRAI generate an illegal instruction exception if
-    imm[5] 6 != 0 but the shift amount is encoded in the lower 6 bits of the I-immediate field for RV64I.
-    """
-
-    rd = result_def(IntRegisterType)
-    rs1 = operand_def(IntRegisterType)
-    immediate = attr_def(IntegerAttr[UI5] | LabelAttr)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        immediate: int | IntegerAttr[UI5] | str | LabelAttr,
-        *,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(immediate, int):
-            immediate = IntegerAttr(immediate, ui5)
-        elif isinstance(immediate, str):
-            immediate = LabelAttr(immediate)
-
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            operands=[rs1],
-            result_types=[rd],
-            attributes={
-                "immediate": immediate,
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.rs1, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["immediate"] = parse_immediate_value(parser, ui5)
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(", ")
-        print_immediate_value(printer, self.immediate)
-        return {"immediate"}
-
-
-class RdRsImmJumpOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations that have one destination register, one source
-    register and one immediate operand.
-
-    This is called I-Type in the RISC-V specification.
-
-    In the RISC-V spec, this is the same as `RdRsImmOperation`. For jumps, the `rd` register
-    is neither an operand, because the stored value is overwritten, nor a result value,
-    because the value in `rd` is not defined after the jump back. So the `rd` makes the
-    most sense as an attribute.
-    """
-
-    rs1 = operand_def(IntRegisterType)
-    rd = opt_attr_def(IntRegisterType)
-    """
-    The rd register here is not a register storing the result, rather the register where
-    the program counter is stored before jumping.
-    """
-    immediate = attr_def(IntegerAttr[SI12] | LabelAttr)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        immediate: int | IntegerAttr[SI12] | str | LabelAttr,
-        *,
-        rd: IntRegisterType | None = None,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(immediate, int):
-            immediate = IntegerAttr(immediate, si12)
-        elif isinstance(immediate, str):
-            immediate = LabelAttr(immediate)
-
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            operands=[rs1],
-            attributes={
-                "immediate": immediate,
-                "rd": rd,
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg | None, ...]:
-        return self.rd, self.rs1, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["immediate"] = parse_immediate_value(parser, si12)
-        if parser.parse_optional_punctuation(","):
-            attributes["rd"] = parser.parse_attribute()
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(", ")
-        print_immediate_value(printer, self.immediate)
-        if self.rd is not None:
-            printer.print_string(", ")
-            printer.print_attribute(self.rd)
-        return {"immediate", "rd"}
-
-
-class RdRsOperation(
-    RISCVCustomFormatOperation, RISCVInstruction, ABC, Generic[RDInvT, RSInvT]
-):
-    """
-    A base class for RISC-V pseudo-instructions that have one destination register and one
-    source register.
-    """
-
-    rd = result_def(RDInvT)
-    rs = operand_def(RSInvT)
-
-    def __init__(
-        self,
-        rs: Operation | SSAValue,
-        *,
-        rd: RDInvT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            operands=[rs],
-            result_types=[rd],
-            attributes={"comment": comment},
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.rs
-
-
-class RdRsIntegerOperation(
-    RdRsOperation[IntRegisterType, RSInvT], ABC, Generic[RSInvT]
-):
-    def __init__(
-        self,
-        rs: Operation | SSAValue,
-        *,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        super().__init__(rs, rd=rd, comment=comment)
-
-
-class RdRsFloatOperation(
-    RdRsOperation[FloatRegisterType, RSInvT], ABC, Generic[RSInvT]
-):
-    def __init__(
-        self,
-        rs: Operation | SSAValue,
-        *,
-        rd: FloatRegisterType = Registers.UNALLOCATED_FLOAT,
-        comment: str | StringAttr | None = None,
-    ):
-        super().__init__(rs, rd=rd, comment=comment)
-
-
-class RsRsOffIntegerOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations that have one source register and a destination
-    register, and an offset.
-
-    This is called B-Type in the RISC-V specification.
-    """
-
-    rs1 = operand_def(IntRegisterType)
-    rs2 = operand_def(IntRegisterType)
-    offset = attr_def(IntegerAttr[SI12] | LabelAttr)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        offset: int | IntegerAttr[SI12] | LabelAttr,
-        *,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(offset, int):
-            offset = IntegerAttr(offset, si12)
-        if isinstance(offset, str):
-            offset = LabelAttr(offset)
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            operands=[rs1, rs2],
-            attributes={
-                "offset": offset,
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rs1, self.rs2, self.offset
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["offset"] = parse_immediate_value(parser, si12)
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(", ")
-        print_immediate_value(printer, self.offset)
-        return {"offset"}
-
-
-class RsRsImmIntegerOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations that have two source registers and an
-    immediate.
-
-    This is called S-Type in the RISC-V specification.
-    """
-
-    rs1 = operand_def(IntRegisterType)
-    rs2 = operand_def(IntRegisterType)
-    immediate = attr_def(IntegerAttr[SI12])
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        immediate: int | IntegerAttr[SI12] | str | LabelAttr,
-        *,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(immediate, int):
-            immediate = IntegerAttr(immediate, si12)
-        elif isinstance(immediate, str):
-            immediate = LabelAttr(immediate)
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            operands=[rs1, rs2],
-            attributes={
-                "immediate": immediate,
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rs1, self.rs2, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["immediate"] = parse_immediate_value(parser, si12)
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(", ")
-        print_immediate_value(printer, self.immediate)
-        return {"immediate"}
-
-
-class RsRsIntegerOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations that have two source
-    registers.
-    """
-
-    rs1 = operand_def(IntRegisterType)
-    rs2 = operand_def(IntRegisterType)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            operands=[rs1, rs2],
-            attributes={
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rs1, self.rs2
-
-
-class NullaryOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations that have neither sources nor destinations.
-    """
-
-    def __init__(
-        self,
-        *,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            attributes={
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return ()
-
-    @classmethod
-    def parse_unresolved_operands(cls, parser: Parser) -> list[UnresolvedOperand]:
-        return []
-
-    def print_op_type(self, printer: Printer) -> None:
-        return
-
-    @classmethod
-    def parse_op_type(
-        cls, parser: Parser
-    ) -> tuple[Sequence[Attribute], Sequence[Attribute]]:
-        return (), ()
-
-
-class CsrReadWriteOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations performing a swap to/from a CSR.
-
-    The 'writeonly' attribute controls the actual behaviour of the operation:
-    * when True, the operation writes the rs value to the CSR but never reads it and
-      in this case rd *must* be allocated to x0
-    * when False, a proper atomic swap is performed and the previous CSR value is
-      returned in rd
-    """
-
-    rd = result_def(IntRegisterType)
-    rs1 = operand_def(IntRegisterType)
-    csr = attr_def(IntegerAttr)
-    writeonly = opt_attr_def(UnitAttr)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        csr: IntegerAttr,
-        *,
-        writeonly: bool = False,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            operands=[rs1],
-            attributes={
-                "csr": csr,
-                "writeonly": UnitAttr() if writeonly else None,
-                "comment": comment,
-            },
-            result_types=[rd],
-        )
-
-    def verify_(self) -> None:
-        if not self.writeonly:
-            return
-        if is_non_zero(self.rd.type):
-            raise VerifyException(
-                "When in 'writeonly' mode, destination must be register x0 (a.k.a. 'zero'), "
-                f"not '{self.rd.type.register_name.data}'"
-            )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.csr, self.rs1
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["csr"] = IntegerAttr(
-            parser.parse_integer(allow_boolean=False, context_msg="Expected csr"),
-            IntegerType(32),
-        )
-        if parser.parse_optional_punctuation(",") is not None:
-            if (flag := parser.parse_str_literal("Expected 'w' flag")) != "w":
-                parser.raise_error(f"Expected 'w' flag, got '{flag}'")
-            attributes["writeonly"] = UnitAttr()
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(", ")
-        self.csr.print_without_type(printer)
-        if self.writeonly is not None:
-            printer.print_string(', "w"')
-        return {"csr", "writeonly"}
-
-
-class CsrBitwiseOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations performing a masked bitwise operation on the
-    CSR while returning the original value.
-
-    The 'readonly' attribute controls the actual behaviour of the operation:
-    * when True, the operation is guaranteed to have no side effects that can
-      be potentially related to writing to a CSR; in this case rs *must be
-      allocated to x0*
-    * when False, the bitwise operations is performed and any side effect related
-      to writing to a CSR takes place even if the mask in rs has no actual bits set.
-    """
-
-    rd = result_def(IntRegisterType)
-    rs1 = operand_def(IntRegisterType)
-    csr = attr_def(IntegerAttr)
-    readonly = opt_attr_def(UnitAttr)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        csr: IntegerAttr,
-        *,
-        readonly: bool = False,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            operands=[rs1],
-            attributes={
-                "csr": csr,
-                "readonly": UnitAttr() if readonly else None,
-                "comment": comment,
-            },
-            result_types=[rd],
-        )
-
-    def verify_(self) -> None:
-        if not self.readonly:
-            return
-        assert isinstance(self.rs1.type, IntRegisterType)
-        if is_non_zero(self.rs1.type):
-            raise VerifyException(
-                "When in 'readonly' mode, source must be register x0 (a.k.a. 'zero'), "
-                f"not '{self.rs1.type.register_name.data}'"
-            )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.csr, self.rs1
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["csr"] = IntegerAttr(
-            parser.parse_integer(allow_boolean=False, context_msg="Expected csr"),
-            IntegerType(32),
-        )
-        if parser.parse_optional_punctuation(",") is not None:
-            if (flag := parser.parse_str_literal("Expected 'r' flag")) != "r":
-                parser.raise_error(f"Expected 'r' flag, got '{flag}'")
-            attributes["readonly"] = UnitAttr()
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(", ")
-        self.csr.print_without_type(printer)
-        if self.readonly is not None:
-            printer.print_string(', "r"')
-        return {"csr", "readonly"}
-
-
-class CsrReadWriteImmOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations performing a write immediate to/read from a CSR.
-
-    The 'writeonly' attribute controls the actual behaviour of the operation:
-    * when True, the operation writes the rs value to the CSR but never reads it and
-      in this case rd *must* be allocated to x0
-    * when False, a proper atomic swap is performed and the previous CSR value is
-      returned in rd
-    """
-
-    rd = result_def(IntRegisterType)
-    csr = attr_def(IntegerAttr)
-    immediate = attr_def(IntegerAttr)
-    writeonly = opt_attr_def(UnitAttr)
-
-    def __init__(
-        self,
-        csr: IntegerAttr,
-        immediate: IntegerAttr,
-        *,
-        writeonly: bool = False,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            attributes={
-                "csr": csr,
-                "immediate": immediate,
-                "writeonly": UnitAttr() if writeonly else None,
-                "comment": comment,
-            },
-            result_types=[rd],
-        )
-
-    def verify_(self) -> None:
-        if self.writeonly is None:
-            return
-        if is_non_zero(self.rd.type):
-            raise VerifyException(
-                "When in 'writeonly' mode, destination must be register x0 (a.k.a. 'zero'), "
-                f"not '{self.rd.type.register_name.data}'"
-            )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg | None, ...]:
-        return self.rd, self.csr, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["csr"] = IntegerAttr(
-            parser.parse_integer(allow_boolean=False, context_msg="Expected csr"),
-            IntegerType(32),
-        )
-        parser.parse_punctuation(",")
-        attributes["immediate"] = parse_immediate_value(parser, IntegerType(32))
-        if parser.parse_optional_punctuation(",") is not None:
-            if (flag := parser.parse_str_literal("Expected 'w' flag")) != "w":
-                parser.raise_error(f"Expected 'w' flag, got '{flag}'")
-            attributes["writeonly"] = UnitAttr()
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(" ")
-        self.csr.print_without_type(printer)
-        printer.print_string(", ")
-        print_immediate_value(printer, self.immediate)
-        if self.writeonly is not None:
-            printer.print_string(', "w"')
-        return {"csr", "immediate", "writeonly"}
-
-
-class CsrBitwiseImmOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RISC-V operations performing a masked bitwise operation on the
-    CSR while returning the original value. The bitmask is specified in the 'immediate'
-    attribute.
-
-    The 'immediate' attribute controls the actual behaviour of the operation:
-    * when equals to zero, the operation is guaranteed to have no side effects
-      that can be potentially related to writing to a CSR;
-    * when not equal to zero, any side effect related to writing to a CSR takes
-      place.
-    """
-
-    rd = result_def(IntRegisterType)
-    csr = attr_def(IntegerAttr)
-    immediate = attr_def(IntegerAttr)
-
-    def __init__(
-        self,
-        csr: IntegerAttr,
-        immediate: IntegerAttr,
-        *,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            attributes={
-                "csr": csr,
-                "immediate": immediate,
-                "comment": comment,
-            },
-            result_types=[rd],
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.csr, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["csr"] = IntegerAttr(
-            parser.parse_integer(allow_boolean=False, context_msg="Expected csr"),
-            IntegerType(32),
-        )
-        parser.parse_punctuation(",")
-        attributes["immediate"] = parse_immediate_value(parser, IntegerType(32))
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(" ")
-        self.csr.print_without_type(printer)
-        printer.print_string(", ")
-        print_immediate_value(printer, self.immediate)
-        return {"csr", "immediate"}
-
-
-# endregion
+from .attrs import (
+    SI20,
+    FastMathFlagsAttr,
+    LabelAttr,
+)
+from .registers import (
+    FloatRegisterType,
+    IntRegisterType,
+    Registers,
+    RISCVRegisterType,
+)
 
 # region RV32I/RV64I: 2.4 Integer Computational Instructions
 
@@ -2441,9 +1096,9 @@ class LwOp(RdRsImmIntegerOperation):
 
     def assembly_line(self) -> str | None:
         instruction_name = self.assembly_instruction_name()
-        value = _assembly_arg_str(self.rd)
-        imm = _assembly_arg_str(self.immediate)
-        offset = _assembly_arg_str(self.rs1)
+        value = assembly_arg_str(self.rd)
+        imm = assembly_arg_str(self.immediate)
+        offset = assembly_arg_str(self.rs1)
         return AssemblyPrinter.assembly_line(
             instruction_name, f"{value}, {imm}({offset})", self.comment
         )
@@ -2508,9 +1163,9 @@ class SwOp(RsRsImmIntegerOperation):
 
     def assembly_line(self) -> str | None:
         instruction_name = self.assembly_instruction_name()
-        value = _assembly_arg_str(self.rs2)
-        imm = _assembly_arg_str(self.immediate)
-        offset = _assembly_arg_str(self.rs1)
+        value = assembly_arg_str(self.rs2)
+        imm = assembly_arg_str(self.immediate)
+        offset = assembly_arg_str(self.rs1)
         return AssemblyPrinter.assembly_line(
             instruction_name, f"{value}, {imm}({offset})", self.comment
         )
@@ -3473,7 +2128,7 @@ class LiOpHasCanonicalizationPatternTrait(HasCanonicalizationPatternsTrait):
 
 
 @irdl_op_definition
-class LiOp(RISCVCustomFormatOperation, RISCVInstruction, ConstantLikeInterface, ABC):
+class LiOp(RISCVCustomFormatOperation, RISCVInstruction, ConstantLikeInterface):
     """
     Loads a 32-bit immediate into rd.
 
@@ -3646,7 +2301,7 @@ class DirectiveOp(
 
     def assembly_line(self) -> str | None:
         if self.value is not None and self.value.data:
-            arg_str = _assembly_arg_str(self.value.data)
+            arg_str = assembly_arg_str(self.value.data)
         else:
             arg_str = ""
 
@@ -3849,61 +2504,6 @@ class WfiOp(NullaryOperation):
 # region RISC-V SSA Helpers
 
 
-class GetAnyRegisterOperation(
-    RISCVCustomFormatOperation,
-    RISCVAsmOperation,
-    RISCVRegallocOperation,
-    ABC,
-    Generic[RDInvT],
-):
-    """
-    This instruction allows us to create an SSAValue with for a given register name. This
-    is useful for bridging the RISC-V convention that stores the result of function calls
-    in `a0` and `a1` into SSA form.
-
-    For example, to generate this assembly:
-    ```
-    jal my_func
-    add a0 s0 a0
-    ```
-
-    One needs to do the following:
-
-    ``` python
-    rhs = riscv.GetRegisterOp(Registers.s0).res
-    riscv.JalOp("my_func")
-    lhs = riscv.GetRegisterOp(Registers.A0).res
-    sum = riscv.AddOp(lhs, rhs, Registers.A0).rd
-    ```
-    """
-
-    res = result_def(RDInvT)
-
-    traits = traits_def(Pure())
-
-    def __init__(
-        self,
-        register_type: RDInvT,
-    ):
-        super().__init__(result_types=[register_type])
-
-    def assembly_line(self) -> str | None:
-        # Don't print assembly for creating a SSA value representing register
-        return None
-
-    @classmethod
-    def parse_op_type(
-        cls, parser: Parser
-    ) -> tuple[Sequence[Attribute], Sequence[Attribute]]:
-        parser.parse_punctuation(":")
-        res_type = parser.parse_attribute()
-        return (), (res_type,)
-
-    def print_op_type(self, printer: Printer) -> None:
-        printer.print_string(" : ")
-        printer.print_attribute(self.res.type)
-
-
 @irdl_op_definition
 class GetRegisterOp(GetAnyRegisterOperation[IntRegisterType]):
     name = "riscv.get_register"
@@ -3970,197 +2570,6 @@ class ParallelMovOp(RISCVRegallocOperation):
 # endregion
 
 # region RV32F: 8 “F” Standard Extension for Single-Precision Floating-Point, Version 2.0
-
-
-class RdRsRsRsFloatOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RV32F operations that take three
-    floating-point input registers and a destination register,
-    e.g: fused-multiply-add (FMA) instructions.
-    """
-
-    rd = result_def(FloatRegisterType)
-    rs1 = operand_def(FloatRegisterType)
-    rs2 = operand_def(FloatRegisterType)
-    rs3 = operand_def(FloatRegisterType)
-
-    traits = traits_def(RegisterAllocatedMemoryEffect())
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        rs3: Operation | SSAValue,
-        *,
-        rd: FloatRegisterType = Registers.UNALLOCATED_FLOAT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            operands=[rs1, rs2, rs3],
-            attributes={
-                "comment": comment,
-            },
-            result_types=[rd],
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.rs1, self.rs2, self.rs3
-
-
-class RdRsRsFloatFloatIntegerOperationWithFastMath(
-    RISCVCustomFormatOperation, RISCVInstruction, ABC
-):
-    """
-    A base class for RISC-V operations that have two source floating-point
-    registers with an integer destination register, and can be annotated with fastmath flags.
-
-    This is called R-Type in the RISC-V specification.
-    """
-
-    rd = result_def(IntRegisterType)
-    rs1 = operand_def(FloatRegisterType)
-    rs2 = operand_def(FloatRegisterType)
-    fastmath = attr_def(FastMathFlagsAttr)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        *,
-        rd: IntRegisterType = Registers.UNALLOCATED_INT,
-        fastmath: FastMathFlagsAttr = FastMathFlagsAttr("none"),
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            operands=[rs1, rs2],
-            attributes={
-                "comment": comment,
-                "fastmath": fastmath,
-            },
-            result_types=[rd],
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.rs1, self.rs2
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        fast = FastMathFlagsAttr("none")
-        if parser.parse_optional_keyword("fastmath") is not None:
-            fast = FastMathFlagsAttr(FastMathFlagsAttr.parse_parameter(parser))
-        attributes["fastmath"] = fast
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        if self.fastmath != FastMathFlagsAttr("none"):
-            printer.print_string(" fastmath")
-            self.fastmath.print_parameter(printer)
-        return {"fastmath"}
-
-
-class RsRsImmFloatOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RV32F operations that have two source registers
-    (one integer and one floating-point) and an immediate.
-    """
-
-    rs1 = operand_def(IntRegisterType)
-    rs2 = operand_def(FloatRegisterType)
-    immediate = attr_def(IntegerAttr[I12])
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        rs2: Operation | SSAValue,
-        immediate: int | IntegerAttr[I12] | str | LabelAttr,
-        *,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(immediate, int):
-            immediate = IntegerAttr(immediate, i12)
-        elif isinstance(immediate, str):
-            immediate = LabelAttr(immediate)
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-
-        super().__init__(
-            operands=[rs1, rs2],
-            attributes={
-                "immediate": immediate,
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rs1, self.rs2, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["immediate"] = parse_immediate_value(parser, i12)
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(", ")
-        print_immediate_value(printer, self.immediate)
-        return {"immediate"}
-
-
-class RdRsImmFloatOperation(RISCVCustomFormatOperation, RISCVInstruction, ABC):
-    """
-    A base class for RV32Foperations that have one floating-point
-    destination register, one source register and
-    one immediate operand.
-    """
-
-    rd = result_def(FloatRegisterType)
-    rs1 = operand_def(IntRegisterType)
-    immediate = attr_def(IntegerAttr[I12] | LabelAttr)
-
-    def __init__(
-        self,
-        rs1: Operation | SSAValue,
-        immediate: int | IntegerAttr[I12] | str | LabelAttr,
-        *,
-        rd: FloatRegisterType = Registers.UNALLOCATED_FLOAT,
-        comment: str | StringAttr | None = None,
-    ):
-        if isinstance(immediate, int):
-            immediate = IntegerAttr(immediate, i12)
-        elif isinstance(immediate, str):
-            immediate = LabelAttr(immediate)
-
-        if isinstance(comment, str):
-            comment = StringAttr(comment)
-        super().__init__(
-            operands=[rs1],
-            result_types=[rd],
-            attributes={
-                "immediate": immediate,
-                "comment": comment,
-            },
-        )
-
-    def assembly_line_args(self) -> tuple[AssemblyInstructionArg, ...]:
-        return self.rd, self.rs1, self.immediate
-
-    @classmethod
-    def custom_parse_attributes(cls, parser: Parser) -> dict[str, Attribute]:
-        attributes = dict[str, Attribute]()
-        attributes["immediate"] = parse_immediate_value(parser, i12)
-        return attributes
-
-    def custom_print_attributes(self, printer: Printer) -> AbstractSet[str]:
-        printer.print_string(", ")
-        print_immediate_value(printer, self.immediate)
-        return {"immediate"}
 
 
 @irdl_op_definition
@@ -4565,9 +2974,9 @@ class FLwOp(RdRsImmFloatOperation):
 
     def assembly_line(self) -> str | None:
         instruction_name = self.assembly_instruction_name()
-        value = _assembly_arg_str(self.rd)
-        imm = _assembly_arg_str(self.immediate)
-        offset = _assembly_arg_str(self.rs1)
+        value = assembly_arg_str(self.rd)
+        imm = assembly_arg_str(self.immediate)
+        offset = assembly_arg_str(self.rs1)
         return AssemblyPrinter.assembly_line(
             instruction_name, f"{value}, {imm}({offset})", self.comment
         )
@@ -4599,9 +3008,9 @@ class FSwOp(RsRsImmFloatOperation):
 
     def assembly_line(self) -> str | None:
         instruction_name = self.assembly_instruction_name()
-        value = _assembly_arg_str(self.rs2)
-        imm = _assembly_arg_str(self.immediate)
-        offset = _assembly_arg_str(self.rs1)
+        value = assembly_arg_str(self.rs2)
+        imm = assembly_arg_str(self.immediate)
+        offset = assembly_arg_str(self.rs1)
         return AssemblyPrinter.assembly_line(
             instruction_name, f"{value}, {imm}({offset})", self.comment
         )
@@ -4803,9 +3212,9 @@ class FLdOp(RdRsImmFloatOperation):
 
     def assembly_line(self) -> str | None:
         instruction_name = self.assembly_instruction_name()
-        value = _assembly_arg_str(self.rd)
-        imm = _assembly_arg_str(self.immediate)
-        offset = _assembly_arg_str(self.rs1)
+        value = assembly_arg_str(self.rd)
+        imm = assembly_arg_str(self.immediate)
+        offset = assembly_arg_str(self.rs1)
         if isinstance(self.immediate, LabelAttr):
             return AssemblyPrinter.assembly_line(
                 instruction_name, f"{value}, {imm}, {offset}", self.comment
@@ -4842,9 +3251,9 @@ class FSdOp(RsRsImmFloatOperation):
 
     def assembly_line(self) -> str | None:
         instruction_name = self.assembly_instruction_name()
-        value = _assembly_arg_str(self.rs2)
-        imm = _assembly_arg_str(self.immediate)
-        offset = _assembly_arg_str(self.rs1)
+        value = assembly_arg_str(self.rs2)
+        imm = assembly_arg_str(self.immediate)
+        offset = assembly_arg_str(self.rs1)
         return AssemblyPrinter.assembly_line(
             instruction_name, f"{value}, {imm}({offset})", self.comment
         )
